@@ -1,41 +1,30 @@
 {
-  "name": "invoicing",
+  "name": "accounting",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@chakra-ui/react": "^1.8.1",
-    "@chakra-ui/theme": "^2.1.0",
-    "@chakra-ui/theme-tools": "^1.3.1",
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@hookform/resolvers": "^2.8.8",
-    "@reduxjs/toolkit": "^1.8.0",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^28.1.4",
-    "@types/node": "^18.0.1",
-    "@types/react": "^18.0.14",
-    "@types/react-dom": "^18.0.5",
-    "@types/react-redux": "^7.1.24",
-    "@types/redux-saga": "^0.10.5",
+    "@chakra-ui/cra-template": "2.0.5",
+    "@chakra-ui/react": "^2.2.6",
+    "@emotion/react": "^11.10.0",
+    "@emotion/styled": "^11.10.0",
+    "@hookform/resolvers": "^2.9.7",
+    "@reduxjs/toolkit": "^1.8.3",
     "big.js": "^6.2.1",
-    "chakra-ui-steps": "^1.6.2",
-    "firebase": "^9.6.7",
-    "framer-motion": "^5.6.0",
+    "chakra-ui-steps": "^1.7.3",
+    "firebase": "^9.9.2",
+    "framer-motion": "^7.0.0",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-datepicker": "^4.8.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^7.34.0",
-    "react-icons": "^4.3.1",
-    "react-redux": "^7.2.6",
-    "react-router": "^6.2.1",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
-    "react-table": "^7.7.0",
-    "recharts": "^2.1.12",
-    "redux": "^4.1.2",
+    "react-icons": "^4.4.0",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "react-table": "^7.8.0",
+    "recharts": "^2.1.13",
+    "redux": "^4.2.0",
     "redux-saga": "^1.1.3",
     "typescript": "^4.7.4",
     "uuid": "^8.3.2",
