@@ -1,37 +1,31 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import { Provider } from "react-redux";
-import store from "./store";
+import { ColorModeScript } from '@chakra-ui/react';
+import React, { StrictMode } from 'react';
+import * as ReactDOM from 'react-dom/client';
+import App from './App';
+import reportWebVitals from './reportWebVitals';
+import * as serviceWorker from './serviceWorker';
 
-import "./index.css";
-import App from "./App";
+import { Provider } from 'react-redux';
+import store from './store';
 
-ReactDOM.render(
-  <React.StrictMode>
+const container = document.getElementById('root');
+const root = ReactDOM.createRoot(container);
+
+root.render(
+  <StrictMode>
+    <ColorModeScript initialColorMode="light" />
     <Provider store={store}>
       <App />
     </Provider>
-  </React.StrictMode>,
-  document.getElementById("root")
+  </StrictMode>
 );
+
+// If you want your app to work offline and load faster, you can change
+// unregister() to register() below. Note this comes with some pitfalls.
+// Learn more about service workers: https://cra.link/PWA
+serviceWorker.unregister();
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-// reportWebVitals();
-
-// "settings": {
-//   "import/resolver": {
-//     "node": {
-//       "paths": [
-//         "src"
-//       ],
-//       "extensions": [
-//         ".js",
-//         ".jsx",
-//         ".ts",
-//         ".tsx"
-//       ]
-//     }
-//   }
-// }
+reportWebVitals();
